requirements.txt
numpy
lxml
black
discord-py
flask
gitpython
google-api-python-client
google-auth-oauthlib
jellyfish
openai
psutil
python-dotenv
slack-sdk
structlog
transformers
ezodf
pydrive
